refactor(onramp): extract src/auth package + preHandler factory (mirrors backend-v2 internal/auth)
